Remove the dynamic table analyzer and function
